+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 3.0.58 (So far...) [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
 
 - Positive adjustment powers now behave better when consumed prior to fading or being manually removed.
+- Healing should now work as expected from 1 source. It can no longer increase above the core value.
 
 ## Version 3.0.58